feat: also mirror the Outlook addon manifest
Generalize the sync script to handle both PostGuard addons. Mirror the
manifest.xml asset from postguard-outlook-addon releases into
static/downloads/postguard-outlook-manifest.xml and update the Outlook
install instructions on /addons to point at the postguard.eu URL
instead of the Azure-hosted manifest.
