+++ src/lib/locales/en.json
@@ -81,7 +81,7 @@
         "title": "Add PostGuard to your mail client",
         "par": "Using the PostGuard addon for Thunderbird or Outlook, you can send encrypted emails in a breeze!",
         "Thunderbird": "<ol><li>Download and save the file <a href=\"/downloads/postguard-tb-addon.xpi\" download><q>postguard-tb-addon.xpi</q></a> somewhere on your computer, where you can find it again easily. </li><li><p>Within Thunderbird:</p><ul><li>under the menu <q>Tools</q>, go to <q>Add-ons and Themes</q>,</li><li> click on the sign, after <q>Manage Your Extensions</q> and select <q>Install Add-on From File ...</q>, </li><li> a file selection window opens; go to the folder where you stored the file <q>postguard-tb-addon.xpi</q> and select it, </li><li> Thunderbird tells you what an add-on can do; when you wish to continue, click the <q>Add</q> button. </li></ul></li><li> Now, when you compose a new mail message, a new header is shown with a button to switch PostGuard on or off. When on, it will encrypt the message for all recipients. Each recipient needs to use the Yivi app to prove that (s)he really posesses the email address that you used. </li></ol>",
-        "Outlook": "<ol><li><p>Adding the add-on:</p><ul><li>Open Outlook</li><li> Under the menu <q>Home</q> or menu item <q>Tools</q>, click on <q>Get Add-ins</q>, </li><li>Click on the left on <q>My add-ins</q>,</li><li> Scroll down to the <q>Custom Addins</q> section. <i>If you do not see this option, your organization probably does not allow custom add-ons to be installed. Please contact your administrator.</i>,</li><li> Click on <q>Add a custom add-in</q>, and then <q>Add from URL...</q>,</li><li> Enter <q>https://irmaseal.z6.web.core.windows.net/manifest.prod.xml</q> and click <q>OK</q> and <q>Install</q>, </li><li>Now the add-in is being installed.</li></ul></li><li> Now, when you compose a new mail message, a new group is visible in the <q>Message</q> menu titled <q>PostGuard Add-in</q>. It contains only one button, called <q>Send encrypted Email</q>. When clicking, it will encrypt and send the message to all recipients. Each recipient needs to use the Yivi app to prove that (s)he really possesses the email address that you used. Also, a copy of the sent mail is stored in the folder <q>PostGuard Sent Items</q> as a draft email (currently, Microsoft only allows to create messages as a draft). </li><li> After receiving a PostGuard encrypted email, the email can be decrypted by clicking on the <q>Decrypt Email</q> button within the <q>PostGuard Add-in</q> group under the <q>Home</q> menu. </li></ol>",
+        "Outlook": "<ol><li><p>Adding the add-on:</p><ul><li>Open Outlook</li><li> Under the menu <q>Home</q> or menu item <q>Tools</q>, click on <q>Get Add-ins</q>, </li><li>Click on the left on <q>My add-ins</q>,</li><li> Scroll down to the <q>Custom Addins</q> section. <i>If you do not see this option, your organization probably does not allow custom add-ons to be installed. Please contact your administrator.</i>,</li><li> Click on <q>Add a custom add-in</q>, and then <q>Add from URL...</q>,</li><li> Enter <q>https://postguard.eu/downloads/postguard-outlook-manifest.xml</q> and click <q>OK</q> and <q>Install</q>, </li><li>Now the add-in is being installed.</li></ul></li><li> Now, when you compose a new mail message, a new group is visible in the <q>Message</q> menu titled <q>PostGuard Add-in</q>. It contains only one button, called <q>Send encrypted Email</q>. When clicking, it will encrypt and send the message to all recipients. Each recipient needs to use the Yivi app to prove that (s)he really possesses the email address that you used. Also, a copy of the sent mail is stored in the folder <q>PostGuard Sent Items</q> as a draft email (currently, Microsoft only allows to create messages as a draft). </li><li> After receiving a PostGuard encrypted email, the email can be decrypted by clicking on the <q>Decrypt Email</q> button within the <q>PostGuard Add-in</q> group under the <q>Home</q> menu. </li></ol>",
         "instruction.header": "Install instructions",
         "instruction.download": "Download for:"
     },
