+++ src/lib/locales/nl.json
@@ -81,7 +81,7 @@
         "title": "Voeg PostGuard toe aan uw e-mailclient",
         "par": "Met de PostGuard add-on voor Thunderbird of Outlook kunt je in een handomdraai versleutelde e-mails verzenden!",
         "Thunderbird": "<ol><li>Download en bewaar het bestand <a href=\"/downloads/postguard-tb-addon.xpi\" download><q>postguard-tb-addon.xpi</q></a> ergens op uw computer, waar je het gemakkelijk terug kunt vinden. </li><li><p>Binnen Thunderbird:</p><ul><li>ga onder het menu <q>Tools</q> naar <q>Add-ons en thema's</q>, </li><li> klik op <q>Uw extensies beheren</q> en selecteer <q>Add-on installeren vanuit bestand...</q>, </li><li> een venster voor bestandsselectie wordt geopend; ga naar de map waar je het bestand <q>postguard-tb-addon.xpi</q> hebt opgeslagen en selecteer het, </li><li> Thunderbird vertelt je wat een add-on kan doen; als je wilt doorgaan, klikt je op de knop <q>Toevoegen</q>. </li></ul></li><li> Als je nu een nieuw mailbericht opstelt, wordt er een nieuwe balk getoond met een knop om PostGuard aan of uit te zetten. Indien ingeschakeld, wordt het bericht voor alle ontvangers versleuteld. Elke ontvanger moet de Yivi-app gebruiken om te bewijzen dat hij/zij echt in het bezit is van het e-mailadres dat je hebt gebruikt. </li></ol>",
-        "Outlook": "<ol><li><p>De add-on toevoegen:</p><ul><li>Open Outlook</li><li> Onder het menu <q>Home</q> of menu-item <q >Tools</q>, klik op <q>Add-ins ophalen</q>, </li><li>Klik links op <q>Mijn add-ins</q>,</li><li> Scrol omlaag naar het gedeelte <q>Aangepaste add-ins</q>. <i>Als je deze optie niet ziet, staat uw organisatie waarschijnlijk niet toe dat aangepaste add-ons worden geïnstalleerd. Neem contact op met uw beheerder.</i>,</li><li> Klik op <q>Een aangepaste Addon toevoegen</q> en vervolgens op <q>Toevoegen vanaf URL...</q>,</li><li> Voer <q>https://irmaseal.z6.web.core.windows.net/manifest.prod.xml</q> in en klik op <q>OK</q> en <q>Installeren</q>, </li><li>Nu wordt de add-in geïnstalleerd.</li></ul></li><li> Als je nu een nieuw e-mailbericht opstelt, is er een nieuwe groep zichtbaar in het menu <q>Message</q> met de naam <q>PostGuard Add- in</q>. Het bevat slechts één knop, genaamd <q>Gecodeerde e-mail verzenden</q>. Wanneer je erop klikt, wordt het bericht gecodeerd en naar alle ontvangers verzonden. Elke ontvanger moet de Yivi-app gebruiken om te bewijzen dat hij/zij echt over het door je gebruikte e-mailadres beschikt. Ook wordt een kopie van de verzonden e-mail opgeslagen in de map <q>PostGuard Sent Items</q> als concept-e-mail (momenteel staat Microsoft alleen toe om berichten als concept te maken). </li><li>Na ontvangst van een met PostGuard versleutelde e-mail, kan de e-mail worden ontsleuteld door te klikken op de knop <q>E-mail ontsleutelen</q> in de groep <q>PostGuard-Addon</q> onder de <q>Start</q>-menu. </li></ol>",
+        "Outlook": "<ol><li><p>De add-on toevoegen:</p><ul><li>Open Outlook</li><li> Onder het menu <q>Home</q> of menu-item <q >Tools</q>, klik op <q>Add-ins ophalen</q>, </li><li>Klik links op <q>Mijn add-ins</q>,</li><li> Scrol omlaag naar het gedeelte <q>Aangepaste add-ins</q>. <i>Als je deze optie niet ziet, staat uw organisatie waarschijnlijk niet toe dat aangepaste add-ons worden geïnstalleerd. Neem contact op met uw beheerder.</i>,</li><li> Klik op <q>Een aangepaste Addon toevoegen</q> en vervolgens op <q>Toevoegen vanaf URL...</q>,</li><li> Voer <q>https://postguard.eu/downloads/postguard-outlook-manifest.xml</q> in en klik op <q>OK</q> en <q>Installeren</q>, </li><li>Nu wordt de add-in geïnstalleerd.</li></ul></li><li> Als je nu een nieuw e-mailbericht opstelt, is er een nieuwe groep zichtbaar in het menu <q>Message</q> met de naam <q>PostGuard Add- in</q>. Het bevat slechts één knop, genaamd <q>Gecodeerde e-mail verzenden</q>. Wanneer je erop klikt, wordt het bericht gecodeerd en naar alle ontvangers verzonden. Elke ontvanger moet de Yivi-app gebruiken om te bewijzen dat hij/zij echt over het door je gebruikte e-mailadres beschikt. Ook wordt een kopie van de verzonden e-mail opgeslagen in de map <q>PostGuard Sent Items</q> als concept-e-mail (momenteel staat Microsoft alleen toe om berichten als concept te maken). </li><li>Na ontvangst van een met PostGuard versleutelde e-mail, kan de e-mail worden ontsleuteld door te klikken op de knop <q>E-mail ontsleutelen</q> in de groep <q>PostGuard-Addon</q> onder de <q>Start</q>-menu. </li></ol>",
         "instruction.header": "Installatie instructies",
         "instruction.download": "Downloaden voor:"
     },
