+++ static/downloads/postguard-outlook-manifest.json
@@ -1,0 +1,9 @@
+{
+  "tag": "v0.1.3",
+  "assetName": "manifest.xml",
+  "sha256": "4646512fab112bd109c8ee56457c7c9676a361f2a36185b78288a4187daaf434",
+  "size": 10357,
+  "publishedAt": "2026-05-06T16:02:32Z",
+  "sourceUrl": "https://github.com/encryption4all/postguard-outlook-addon/releases/download/v0.1.3/manifest.xml",
+  "releaseUrl": "https://github.com/encryption4all/postguard-outlook-addon/releases/tag/v0.1.3"
+}
